event service & create event & create event session & speaker - api

+++ Services/EventService/EventSphere.EventService.Domain/Entities/EventSession.cs
@@ -5,9 +5,9 @@ namespace EventSphere.EventService.Domain.Entities
     public class EventSession : BaseEntity
     {
         public int RecordId { get; set; }
-        public long EventId { get; set; }
+        public int EventId { get; set; }
         public Event Event { get; set; }
-        public long SpeakerId { get; set; }
+        public int SpeakerId { get; set; }
         public Speaker Speaker { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
